Componetize SelectiveColumnReader
Split SelectiveColumnReader into format-specific readers and extract
generic logic for value processing,filtering and the column visitor
pattern.

The objective is reusing common pieces in selective readers for Parquet and Alpha.

The next step is adapting the Visitor signatures so that a single
Visitor can handle alternating direct and dictionary encodings in the
same stripe.
